1601610 bug while editing integer/long columns

+++ sql12/fw/src/net/sourceforge/squirrel_sql/fw/datasetviewer/cellcomponent/DataTypeLong.java
@@ -411,7 +411,7 @@ public class DataTypeLong
 			pstmt.setNull(position, _colDef.getSqlType());
 		}
 		else {
-			pstmt.setInt(position, ((Long)value).intValue());
+			pstmt.setLong(position, ((Long)value).longValue());
 		}
 	}
